analyzing_using_pyfluent2/requirements.txt
# Author: Jiwoo Lee
# PyFluent CFD Analysis Requirements

# Core PyFluent
ansys-fluent-core==0.32.1

# PyVista for VTP export and visualization
pyvista>=0.44.0
vtk>=9.2.0

# Mesh processing
trimesh>=3.23.0
meshio>=5.3.0
scipy>=1.11.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0
json5>=0.9.0

# Progress tracking
tqdm>=4.66.0

# Logging and utilities
colorlog>=6.7.0
pathlib
argparse 